+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Welcome modal now relies on installationId to determine whether to show
 - Add Requirements to README, and "engines" field to package.json
 - Add Segmented Plugin
+- Replace WebView search bar with Search Bar Plugin
 
 ## v0.15.0
 - [iOS] Remove update bundle version build phase